JWT feature implemented

+++ routes/channels.js
@@ -1,30 +1,60 @@
 const express = require('express');
 const router = express.Router();
-const { param, body, validationResult } = require('express-validator');
-const { connection } = require('../db');
+const { param, body, validationResult, cookie } = require('express-validator');
+const { connection, pool } = require('../db');
+const { verify } = require('../jwt');
+const { getUserById } = require('../services/user-service');
 
 
 /* Utility Functions */
 const validate = (req, res, next) => {
     const errors = validationResult(req);
-    if (!errors.isEmpty()) {
+    if (errors.isEmpty() === false) {
         return res.status(400).json(errors.array());
     }
 
     next();
 }
 
-function send400(res, msg) {
+function send400(res, msg = 'An error occurred. Please try again.') {
+    if (typeof msg === 'object') {
+        return res.status(400).json(msg);
+    }
+
     res.status(400).json({
         error: msg
     });
 }
 
-function send404(res, msg) {
+function send404(res, msg = 'Resource not found.') {
+    if (typeof msg === 'object') {
+        return res.status(404).json(msg);
+    }
+    
     res.status(404).json({
         error: msg
     });
 }
+
+/* Middleware */
+router.use((req, res, next) => {
+    const token = req.cookies.token;
+    if (!token) {
+        return res.status(401).json({
+            error: 'Unauthorized'
+        });
+    }
+
+    const payload = verify(token);
+    if (!payload) {
+        return res.status(401).json({
+            error: 'Unauthorized'
+        });
+    }
+
+    req.user = payload;
+    next();
+});
 
 /* API Routes */
 router.route('/:id')
@@ -37,6 +67,13 @@ router.route('/:id')
         ],
         (req, res) => {
             let { id } = req.params;
+            const { userId } = req.user;
+
+            if (id !== userId) {
+                return res.status(403).json({
+                    error: 'You are not authorized to perform this action.'
+                });
+            }
 
             // Check if channel name is provided
             const { channelName } = req.body;
@@ -72,6 +109,13 @@ router.route('/:id')
         ],        
         (req, res) => {
             let { id } = req.params;
+            const { userId } = req.user;
+
+            if (id !== userId) {
+                return res.status(403).json({
+                    error: 'You are not authorized to perform this action.'
+                });
+            }
 
             const query = 'DELETE FROM channels WHERE id = ?';
             connection.query(
@@ -128,48 +172,42 @@ router.route('/')
     // Create a new channel
     .post(
         [
-            body('userId').notEmpty().isInt().withMessage('Provide a valid user id.'),
             body('channelName').notEmpty().isString().withMessage('Provide a valid channel name.'),
             validate
         ],
-        (req, res) => {
-            const { userId, channelName } = req.body;
+        async (req, res) => {
+            const { channelName } = req.body;
+            const { sub } = req.user;
+
+            const user = await getUserById(sub);
+            if (!user) {
+                return send404(res, 'User not found.');
+            }
 
             const query = 'INSERT INTO channels (user_id, name) VALUES (?, ?)';
-            connection.query(
-                query,
-                [userId, channelName],
-                (error, results, fields) => {
-                    if (error) {
-                        return send400(res, 'An error occurred. Please try again.');
-                    }
+            const values = [sub, channelName];
 
-                    if (results.affectedRows > 0) {
-                        return res.status(201).json({
-                            message: 'Channel has been created.'
-                        });
-                    }
-
-                    return res.status(400).json({
-                        error: 'Channel could not be created.'
+            try {
+                const queryResult = await pool.query(query, values);
+                if (queryResult[0].affectedRows > 0) {
+                    return res.status(201).json({
+                        message: 'Channel has been created.'
                     });
                 }
-            );
+            } catch (error) {
+                return send400(res, 'An error occurred. Please try again.');
+            }
     })
     // Get all channels for a user
     .get(
-        [
-            body('userId').notEmpty().isInt().withMessage('Provide a valid user id.'),
-            validate
-        ], 
         (req, res) => {
-            const { userId } = req.body;
+            const { sub } = req.user;
 
             // Check if there are channels
             const query = 'SELECT * FROM channels WHERE user_id = ?';
             connection.query(
                 query,
-                [userId],
+                [sub],
                 (error, results, fields) => {
                     if (error) {
                         return send400(res, 'An error occurred. Please try again.');
